elyune-backend/requirements.txt
asgiref==3.11.0
django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
sqlparse==0.5.5
psycopg2-binary==2.9.10
gunicorn==23.0.0
python-decouple==3.8
boto3==1.35.95
django-storages==1.14.4
celery==5.4.0
redis==5.2.1
flower==2.0.1
deepgram-sdk==3.10.0
google-generativeai==0.8.3
ffmpeg-python==0.2.0
djangorestframework-simplejwt==5.4.0
django-filter==24.3
